Add files via upload
I am excited to announce my latest project, Alternative, a decentralized application (Dapp) that solves a major problem faced by users when sending Ether from one wallet to another. Typically, users struggle to remember the unique address of the recipient and often forget why they sent the Ether in the first place. Alternative addresses this challenge by allowing users to send Ether by providing the recipient's unique username.

Alternative is built on a smart contract that enables users to log in with their unique name, which is then mapped to their Ethereum wallet address. Once a user has logged in, they can send Ether to another user by simply entering the recipient's username and a message to accompany the transaction. The smart contract captures all transaction details, and the sent and received data can be viewed on the Alternative Dapp.

To ensure a seamless user experience, please note that when you try to log in to Alternative and the login does not work, it could be because you are using a username that is already being used by another user. Please remember to use a unique username to avoid this issue.

Additionally, if you are trying to use Alternative from a smartphone, please make sure to use the Metamask explorer browser for optimal performance.

The simplicity and convenience of Alternative make it a game-changer in the crypto space, eliminating the need to remember lengthy wallet addresses and offering a more personalized approach to transactions. The Dapp has been deployed on the Sepolia test network and can be accessed through the Metamask browser extension or app.

To get started with Alternative, download Metamask, connect to the Sepolia test network, and log in with a unique username. You can then send Ether to other users by entering their username and accompanying message. Please note that the recipient must be logged in to the Dapp and must have provided their username during the signup process.

I invite you to try out Alternative and experience the future of decentralized transactions. Check out the Dapp at https://usman-alternative.netlify.app/.